\name{SWAT2R-package}
\Rdversion{1.1}
\alias{SWAT2R-package}
\alias{SWAT2R}
\docType{package}
\title{
SWAT2R
}
\description{
Package for reading the main output files of the Soil & Water Assessment Tool hydrological model (SWAT, http://www.brc.tamus.edu/swat/) and for creating some inputs to SWAT-CUP (http://www.eawag.ch/forschung/siam/software/swat/). Reading functions are particulary efficient for large output files (hundreds of Mb)
License
}
\details{
\tabular{ll}{
Package: \tab SWAT2R\cr
Type: \tab Package\cr
Version: \tab 0.1.5-2\cr
Date: \tab 2012-04-16\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Mauricio Zambrano Bigiarini <mzb.devel@gmail.com>

Maintainer: Mauricio Zambrano Bigiarini <mzb.devel@gmail.com>
}
\references{
Arnold, J. G., R. Srinivasan, R. S. Muttiah, and J. R.Williams (1998), Large area hydrologic modeling and assessment part I: Model development, Journal of the American Water Resources Association, 34(1), 73-89 \cr

Arnold, J. G., and N. Fohrer (2005), SWAT2000: current capabilities and research opportunities in applied watershed modelling, Hydrological Processes, 19(3), 563-572, doi:10.1002/hyp.5611 \cr

Neitsch, S. L., J. G. Arnold, J. R. Kiniry, and J. R. Williams (2005a), Soil and Water Assessment Tool Theoretical Documentation Version 2005, Grassland, Soil and Water Research Laboratory; Agricultural Research Service 808 East Blackland Road; Temple, Texas 76502; Blackland Research \& Extension Center; Texas Agricultural Experiment Station 720 East Blackland Road Temple, Texas 76502, USA \cr

Neitsch, S. L., J. G. Arnold, J. R. Kiniry, R. Srinivasan, and J. R. Williams (2005b), Soil and Water Assessment Tool Input/Output file documentation, Version 2005, Grassland, Soil andWater Research Laboratory; Agricultural Research Service 808 East Blackland Road; Temple, Texas 76502; Blackland Research \& Extension Center; Texas Agricultural Experiment Station 720 East Blackland Road Temple, Texas 76502, USA. 
}
%%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
\code{\link[hydroTSM:hydroTSM]{hydroTSM}}
}
%%\examples{
%%~~ simple examples of the most important functions ~~
%%}
